+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # 13.7.8 - "Twin Shields"
 - small improvements for challenge GM features during entering data
+- import challenges via special JSON files
+- set adventure, origin and greatness might 
+- improvements here and there regarding challenge rendering
 
 # 13.7.7 - "Shield Bash"
 - added might for tags, right click on a tag and you can turn it into a mighty tag